delete special_polys.rs
