+++ jolt-core/src/r1cs/mod.rs
@@ -5,4 +5,3 @@ pub mod constraints;
 pub mod key;
 pub mod ops;
 pub mod spartan;
-pub mod special_polys;
